Age fix

+++ Numpuk2/Numpuk2.Queries/Models/ClientResponse.cs
@@ -6,7 +6,7 @@ namespace Numpuk2.Queries.Models
     public class ClientResponse
     {
         public string Id { get; set; }
-        public int Age { get; set; }
+        public double Age { get; set; }
         public Gender Gender { get; set; }
     }
 }
